pyproject.toml
[tool.poetry]
name = "coursework-5"
version = "0.1.0"
description = ""
authors = ["Alexandr Lebedev <alexandr.lebedev.sp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.dev.dependencies]
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
